refactor attempt 2
